filter dcas output file by permitted countries to view
