+++ django_project/frontend/api_views/__init__.py
@@ -1,2 +1,3 @@
 from frontend.api_views.auth import *  # noqa
 from frontend.api_views.api_keys import *  # noqa
+from frontend.api_views.dcas import *  # noqa
