+++ django_project/frontend/views.py
@@ -2,7 +2,6 @@
 
 import json
 from urllib.parse import urlparse
-from datetime import timedelta
 
 import requests
 from django.conf import settings
@@ -11,17 +10,11 @@ from django.utils.decorators import method_decorator
 from django.views.decorators.csrf import csrf_exempt
 from django.views.generic import TemplateView, View
 from django.urls import reverse
-from django.utils import timezone
-from django.core.files.storage import storages
-from rest_framework import generics, permissions
+from rest_framework import permissions
 from rest_framework.views import APIView
 from rest_framework.response import Response
 
 
-from frontend.permissions import IsKalroUser
-from dcas.models.output import DCASOutput
-from dcas.models.download_log import DCASDownloadLog
-from frontend.serializers import OutputSerializer
 from frontend.models import PagePermission
 from gap.models.preferences import Preferences
 
@@ -136,52 +129,6 @@ class EmailCheckView(TemplateView):
         context = get_base_context(context)
 
         return context
-
-
-class OutputListView(generics.ListAPIView):
-    """Return recent CSV outputs (last 2 weeks) for KALRO users/admins."""
-
-    permission_classes = [permissions.IsAuthenticated, IsKalroUser]
-    serializer_class = OutputSerializer
-    pagination_class = None  # client-side pagination in React
-
-    def get_queryset(self):
-        """Return queryset of recent DCAS outputs."""
-        cutoff = timezone.now() - timedelta(weeks=2)
-        return (
-            DCASOutput.objects.filter(
-                file_name__iendswith=".csv", delivered_at__gte=cutoff
-            ).order_by("-delivered_at")
-        )
-
-
-class OutputDownloadView(APIView):
-    """View to generate presigned URL for downloading DCAS output files."""
-
-    permission_classes = [permissions.IsAuthenticated, IsKalroUser]
-
-    def get(self, request, pk: int, *args, **kwargs):
-        """Generate a presigned URL for downloading a DCAS output file."""
-        output = generics.get_object_or_404(DCASOutput, pk=pk)
-        storage = storages["gap_products"]
-
-        presigned_url = storage.url(
-            name=output.path,
-            expire=900,
-            parameters={
-                "ResponseContentDisposition":
-                f'attachment; filename="{output.file_name}"',
-            },
-        )
-        presigned = presigned_url
-        # **new** – persist audit row
-        DCASDownloadLog.objects.create(output=output, user=request.user)
-        if settings.DEBUG:
-            presigned = presigned_url.replace(
-                "http://minio:9000", "http://localhost:9010"
-            )
-
-        return Response({"url": presigned})
 
 
 class PermittedPagesView(APIView):
